Reorganized components structure
