+++ react-map-gl-amplify-here-map/src/components/overlays/Pin.js
