+++ react-map-gl-amplify-here-map/src/components/primitives/Slider.js
