+++ react-map-gl-amplify-here-map/src/components/primitives/Switch.js
