+++ react-map-gl-amplify-here-map/src/hooks/useAmazonLocationService.js
