+++ react-map-gl-amplify-here-map/src/hooks/useGeolocation.js
