+++ react-map-gl-amplify-here-map/src/hooks/useWindowSize.js
